Delete unnecessary cell
